Convert scripts to tests
